Parses all Instagram accounts featured on NUSSync
This algorithm looks through all Instagram accounts in the JSON returned by the NUSSync organisations endpoint. It then retrieves the latest 5 posts from those accounts.

The JSON endpoint stores the accounts as a Instagram link. The algorithm retrieves the username from this link and searches for the userID through the topsearch API, which it then passes to the existing createInstaEvent function.

createInstaEvents in index.js:
Now takes in organisation name which is stored in database. This forms the basis for the document id as well.

Issues:
At the moment, the topsearch API does not work on cloud function. Hence, we are only able to run allInsta locally.

Others:
Time is now stored as timestamp in event document, allowing documents to be sorted chronologically.

EventDetailFragment now has a delete event button.