Allows event posts to be created and edited.
EditButton in UpdatesPagerAdapter uses the default onItemSelected listener, which listens to the edit button only instead of taps on the post itself.
The onItemSelected method is implemented in EventDetailFragment to navigate to EditPostFragment.

Date of event post cannot be updated and reflects current date of creation.

Issues:
Event editing (including post updates) does not check that current user is organising user. [This is to help in testing as events are from Instagram and we do not have email of the clubs]